To clear noise, instrument for non-kani in macros.
